requirements.txt
flask
tiktoken
trafilatura
PyPDF2
requests
langchain-google-genai
langchain-tavily
python-dotenv